qb cont

+++ x/tests/loop2.c
@@ -1,6 +1,6 @@
 void f(){
  int32_t i;
- i=(int32_t)(0);
+ i=0;
  do{
   i=(i*2);
   i++;
